providers/arangodb/pyproject.toml
[build-system]
requires = ["flit_core >=3.8.0,<4"]
build-backend = "flit.buildapi"

[project]
name = "apache-airflow-providers-arangodb"
version = "2.1.0"
authors = [
    {name = "Apache Airflow PMC", email = "dev@airflow.apache.org"},
]
maintainers = [
    {name = "Apache Airflow PMC", email = "dev@airflow.apache.org"},
]
description = "Provider for Apache Airflow. Implements apache-airflow-providers-arangodb package"
readme = "README.rst"
requires-python = ">=3.7"
keywords = ["airflow", "airflow-provider", "apache-airflow-providers-arangodb"]
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: System :: Monitoring",
]


dependencies = [
    "apache-airflow>=2.3.0",
    "python-arango>=7.3.2",
]

[project.urls]
"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-arangodb/2.1.0/"
"Bug Tracker" = "https://github.com/apache/airflow/issues"
"Source Code" = "https://github.com/apache/airflow"
"Slack Chat" = "https://s.apache.org/airflow-slack"
"Twitter" = "https://twitter.com/ApacheAirflow"
"YouTube" = "https://www.youtube.com/channel/UCSXwxpWZQ7XZ1WL3wqevChA/"

[project.entry-points."apache_airflow_provider"]
provider_info = "airflow.providers.arangodb.get_provider_info:get_provider_info"

[tool.flit.module]
name = "airflow.providers.arangodb"